Improve performance.
CocoaTouch was spending a huge amount of time (approximately 5000ms) solving the autolayout constraints for the product summary cells. The performance problem persisted even after simplifying the constraints, so this may be a bug in the autolayout engine.

I worked around the problem by switching to old-style autosizing. If we need complex layout in the future, we may need to build and layout the view in code.