Don't restore like list.
The list can change between restore state being saved and being
loaded if the app crashes or is killed in the debugger. This results
in 404 errors when requesting the likes in the list. The like list
and selected retailer URLs do not need to be persisted because they
can always be restored from the user resource.